Automobile 1.1.3
This Program meets and exceeds the requirements for the first Assignment for De Anza Advanced Programming course. It begins to implement the required features for 1.2 Which assignment 2. The last submission is very close to this one. Minor changes we're made specifically to how the arrays were handling data into more organized arrayLists. The objective is for each class to eventually stand alone in it's capability to read any data for this auto project. I.E. handle any auto object such as a motor bike, a motor boat, a military vehicle, aircraft, and finally cars.

Next
Interfaces, GENERICS, GENERICS, GENERICS. Exception handling, Documentation, and clean UI/Code(...sometimes the hardest part).