AutoJava 2.7.5
Arbitrary version numbers aside. I am so happy to have my program come this far.
This program meets the requirements of Assignment 2 , except for that it lacks Exception handling.
It almost meets the requirements of Assignment 3, except for that is lacks the API for changing Options specifically.

The program in its present state, has no real main for execution, although it contains all methods for testing. This commit is very raw, where as the next commit today will be much more refined. Including the needed changes for a real applet.